+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 0.18.14
+
 ## 0.18.13
 
 * Catch rare condition in Increment Version command where no symbols are found in info.json